requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.4
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
celery==5.3.4
nltk==3.8.1
pandas==2.1.1
numpy==1.26.0
pyjwt==2.8.0
gunicorn==21.2.0
whitenoise==6.5.0 